namma-ai/requirements.txt
google-generativeai
langchain
langchain-openai
langchain-community
langchain-huggingface
chromadb
duckduckgo-search
python-dotenv
PyMuPDF
tqdm
google-search-results
beautifulsoup4
requests
